+++ client/packages/host/src/components/Navigation/CatalogueNav.tsx
@@ -7,9 +7,7 @@ import {
   AppNavLink,
   ListIcon,
   AppNavSection,
-  useIsCentralServerApi,
 } from '@openmsupply-client/common';
-import { Environment } from '@openmsupply-client/config';
 import { AppRoute } from '@openmsupply-client/config';
 import { useNestedNav } from './useNestedNav';
 
@@ -18,7 +16,6 @@ export const CatalogueNav: FC = () => {
     RouteBuilder.create(AppRoute.Catalogue).addWildCard().build()
   );
   const t = useTranslation('app');
-  const isCentralServer = useIsCentralServerApi();
 
   return (
     <AppNavSection isActive={isActive} to={AppRoute.Catalogue}>
@@ -51,14 +48,6 @@ export const CatalogueNav: FC = () => {
               .addPart(AppRoute.MasterLists)
               .build()}
             text={t('master-lists')}
-          />
-          <AppNavLink
-            visible={isCentralServer && Environment.FEATURE_GAPS}
-            end
-            to={RouteBuilder.create(AppRoute.Catalogue)
-              .addPart(AppRoute.IndicatorsDemographics)
-              .build()}
-            text={t('indicators-demographics')}
           />
         </List>
       </Collapse>
